feat(tui): expand git view details
