+++ src/tui/mod.rs
@@ -31,7 +31,9 @@ pub mod color_palette;
 pub mod git_branch;
 pub mod git_capture;
 pub mod git_diff;
+pub mod git_lines;
 pub mod git_log;
+pub mod git_style;
 pub mod git_view;
 pub mod message_formatter;
 pub mod ralph_view;
